+++ R/bpipe.R
@@ -101,11 +101,15 @@ apply_left.default <- function(pipe_left_arg,
        enclos = pipe_environment)
 }
 
+
+
+
+
 #' S3 dispatch on class of pipe_right_argument.
 #'
 #' Triggered if right hand side was a name that does not resolve to a function.
 #' For formal documentation please see \url{https://github.com/WinVector/wrapr/blob/master/extras/wrapr_pipe.pdf}.
-#' Since both items are now values we could like to S4 dispatch at this point, but
+#' Since both items are now values we might like to S4 dispatch at this point, but
 #' this path is only tiggered with the right argument starts name-line so most
 #' uses would violate referential transparency for a small return.
 #'
@@ -147,6 +151,11 @@ apply_right <- function(pipe_left_arg,
   UseMethod("apply_right", pipe_right_arg)
 }
 
+methods::setGeneric(
+  name = "apply_right",
+  def = apply_right,
+  signature = c("pipe_left_arg", "pipe_right_arg"),
+  simpleInheritanceOnly = FALSE)
 
 #' S3 dispatch on type of pipe_right_argument.
 #'