+++ NEWS.md
@@ -2,6 +2,7 @@
 # wrapr 1.5.2 2018-07-11
  
  * Documentation fixes.
+ * S4 dispatch apply_right.
  * split based partition_tables()
 
 # wrapr 1.5.1 2018-07-07